Started building login
Currently, the user can login and invalid credentials are not allowed as the authentication is done on the server side. If failed, an alert is shown. If succeeded, the token is stored to flutter_secure_storage.

NOTE: In order to use flutter_secure_storage, the minimun Android SDK was raised to 18.